refactor constants, helper functions and imports to reduce UI bundle size
