+++ src/controllers/portfolio/portfolio.ts
@@ -55,7 +55,6 @@ import {
   convertApiTokenDataToTokenDataCache,
   erc721CollectionToLearnedAssetKeys,
   formatExternalHintsAPIResponse,
-  getFlags,
   getHintsError,
   getSpecialHints,
   getTotal,
@@ -91,6 +90,7 @@ import {
   TokenValidationResult
 } from '../../libs/portfolio/interfaces'
 import { PORTFOLIO_LIB_ERROR_NAMES } from '../../libs/portfolio/portfolio'
+import { getFlags } from '../../libs/portfolio/tokenProcessing'
 import { BindedRelayerCall, relayerCall } from '../../libs/relayerCall/relayerCall'
 import { isInternalChain } from '../../libs/selectedAccount/selectedAccount'
 import EventEmitter from '../eventEmitter/eventEmitter'
