+++ src/controllers/safe/safe.ts
@@ -1,11 +1,6 @@
 import { toBeHex } from 'ethers'
 
-import SafeApiKit, {
-  SafeCreationInfoResponse,
-  SafeInfoResponse,
-  SafeMessage
-} from '@safe-global/api-kit'
-import { SafeMultisigConfirmationResponse } from '@safe-global/types-kit'
+import SafeApiKit from '@safe-global/api-kit'
 
 import { FETCH_SAFE_TXNS } from '../../consts/intervals'
 import { SAFE_NETWORKS, safeNullOwner } from '../../consts/safe'
@@ -24,6 +19,9 @@ import {
   SafeResults
 } from '../../libs/safe/safe'
 import EventEmitter from '../eventEmitter/eventEmitter'
+
+import type { SafeCreationInfoResponse, SafeInfoResponse, SafeMessage } from '@safe-global/api-kit'
+import type { SafeMultisigConfirmationResponse } from '@safe-global/types-kit'
 
 export const STATUS_WRAPPED_METHODS = {
   findSafe: 'INITIAL'
